fastapi-publisher/requirements.txt
xenarch[fastapi]
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv
